fix sql migration

+++ hedera-mirror-importer/src/main/resources/db/migration/v2/V2.4.1__redistribute_transaction_hash.sql
@@ -2,10 +2,10 @@ drop table if exists transaction_hash;
 
 create table if not exists transaction_hash
 (
-    consensus_timestamp bigint not null,
-    distribution_id     bigint not null,
-    hash                bytea  not null,
-    payer_account_id    bigint not null
+    consensus_timestamp bigint   not null,
+    distribution_id     smallint not null,
+    hash                bytea    not null,
+    payer_account_id    bigint   not null
 );
 comment on table transaction_hash is 'Network transaction hash to consensus timestamp mapping';
